requirements.txt
google-cloud-bigquery==3.41.0
google-auth
gspread
pandas==2.3.3
numpy
streamlit
plotly
unidecode
db-dtypes
pyarrow
pandas-gbq
python-dotenv
openpyxl
